Read available game balance from merkle tree & optimize overview page
