+++ apps/app/src/hooks/useComicsBalance.ts
@@ -1,6 +1,6 @@
 'use client';
 
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import type { AddressLike } from 'ethers6';
 import type { Contracts } from '@/types/web3';
 import type { Comic } from '@/types/marketplace';
@@ -28,18 +28,12 @@ export default function useComicsBalance(
   const [loading, setLoading] = useState(true);
   const [comicsBalance, setComicsBal] = useState<Comic[]>([]);
   const marketplaceContract = useMemo(() => imxContracts[MARKETPLACE_CONTRACT], [imxContracts]);
-  const firstRenderRef = useRef(true);
 
   useEffect(() => {
     async function checkUserComics() {
       const ownerArr = [address, address, address, address, address, address] as AddressLike[];
       const comicIds = [1, 2, 3, 4, 5, 6];
-      console.log('Fetching Comics for', address);
       const comicsData = await marketplaceContract.balanceOfBatch(ownerArr, comicIds);
-      console.log(
-        'Comics Data',
-        comicsData.map((c: bigint) => c.toString()),
-      );
 
       if (comicsData.some((c: bigint) => c > 0)) {
         setComicsBal(
@@ -53,10 +47,6 @@ export default function useComicsBalance(
     }
 
     if (address && marketplaceContract) {
-      // if (firstRenderRef.current) {
-      //   firstRenderRef.current = false;
-      //   return;
-      // }
       // eslint-disable-next-line no-void
       void checkUserComics();
     }
